fix(hubble-stats): Build error caused by refactoring
