+++ libs/webb-ui-components/src/next-utils/index.ts
@@ -1,1 +1,1 @@
-export * from './sidebarState';
+export * from './sidebarState2';
